REFACTOR: Split hard skills into different files - Makes it easier to manage - Will make it easier to add relations
