+++ database/skills/skills.ts
@@ -1,11 +1,39 @@
 import Skill from "@/types/skills";
 import softSkills from "./softSkills";
-import hardSkills from "./hardSkills";
 import generalSkills from "./generalSkills";
+import technicalHardSkillsFullStackWebDev from "./technicalHardSkills/technicalHardSkillsFullStackWebDev";
+import technicalHardSkillsAPIs from "./technicalHardSkills/technicalHardSkillsAPIs";
+import technicalHardSkillsBackendWebDev from "./technicalHardSkills/technicalHardSkillsBackendWebDev";
+import technicalHardSkillsCloudComputing from "./technicalHardSkills/technicalHardSkillsCloudComputing";
+import technicalHardSkillsDatabases from "./technicalHardSkills/technicalHardSkillsDatabases";
+import technicalHardSkillsDevOps from "./technicalHardSkills/technicalHardSkillsDevOps";
+import technicalHardSkillsFrontendWebDev from "./technicalHardSkills/technicalHardSkillsFrontendWebDev";
+import technicalHardSkillsMaths from "./technicalHardSkills/technicalHardSkillsMaths";
+import technicalHardSkillsMLDS from "./technicalHardSkills/technicalHardSkillsMLDS";
+import technicalHardSkillsOthers from "./technicalHardSkills/technicalHardSkillsOthers";
+import technicalHardSkillsProjectManagers from "./technicalHardSkills/technicalHardSkillsProjectManagers";
+import technicalHardSkillsTesting from "./technicalHardSkills/technicalHardSkillsTesting";
+import technicalHardSkillsVCS from "./technicalHardSkills/technicalHardSkillsVCS";
 
-const allSkills: Skill[] = [...hardSkills, ...generalSkills, ...softSkills];
+const allHardSkills: Skill[] = [
+  ...technicalHardSkillsFullStackWebDev,
+  ...technicalHardSkillsAPIs,
+  ...technicalHardSkillsBackendWebDev,
+  ...technicalHardSkillsCloudComputing,
+  ...technicalHardSkillsDatabases,
+  ...technicalHardSkillsDevOps,
+  ...technicalHardSkillsFrontendWebDev,
+  ...technicalHardSkillsMaths,
+  ...technicalHardSkillsMLDS,
+  ...technicalHardSkillsOthers,
+  ...technicalHardSkillsProjectManagers,
+  ...technicalHardSkillsTesting,
+  ...technicalHardSkillsVCS,
+];
 
-export const technologies: Skill[] = hardSkills.filter(
+const allSkills: Skill[] = [...allHardSkills, ...generalSkills, ...softSkills];
+
+export const technologies: Skill[] = allHardSkills.filter(
   (skill) => skill.isMainSkill
 );
