+++ database/skills/technicalHardSkills/technicalHardSkillsMaths.ts
@@ -1,0 +1,117 @@
+import Skill from "@/types/skills";
+import { mathematics } from "../generalSkills";
+
+export const probability: Skill = {
+  name: "Probability",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "probability",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const statistics: Skill = {
+  name: "Statistics",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "statistics",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const calculus: Skill = {
+  name: "Calculus",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "calculus",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const trigonometry: Skill = {
+  name: "Trigonometry",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "trigonomtry",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const algebra: Skill = {
+  name: "Algebra",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "algebra",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const linearAlgebra: Skill = {
+  name: "Linear Algebra",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "linear-algebra",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const discrete: Skill = {
+  name: "Discrete",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "discrete-mathematics",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const geometry: Skill = {
+  name: "Geometry",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "geometry",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const logics: Skill = {
+  name: "Logics",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "logics",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const vectors: Skill = {
+  name: "Vectors",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "vectors",
+  technicalGeneralSkills: [mathematics],
+};
+
+export const mechanics: Skill = {
+  name: "Mechanics",
+  category: "Mathematics",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "mechanics",
+  technicalGeneralSkills: [mathematics],
+};
+
+const technicalHardSkillsMaths: Skill[] = [
+  probability,
+  statistics,
+  calculus,
+  trigonometry,
+  algebra,
+  linearAlgebra,
+  discrete,
+  geometry,
+  logics,
+  vectors,
+  mechanics,
+];
+
+export default technicalHardSkillsMaths;
