+++ database/skills/technicalHardSkills/technicalHardSkillsProjectManagers.ts
@@ -1,0 +1,85 @@
+import Skill from "@/types/skills";
+
+export const npm: Skill = {
+  name: "NPM",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "npm",
+};
+
+export const yarn: Skill = {
+  name: "Yarn",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "yarn",
+};
+
+export const pnpm: Skill = {
+  name: "PNPM",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "pnpm",
+};
+
+export const poetry: Skill = {
+  name: "Poetry",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "poetry-python",
+};
+
+export const pybuilder: Skill = {
+  name: "PyBuilder",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "pybuilder",
+};
+
+export const maven: Skill = {
+  name: "Maven",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "maven",
+};
+
+export const gradle: Skill = {
+  name: "Gradle",
+  category: "Project Managers",
+  isMainSkill: true,
+  skillType: "hard",
+  slug: "gradle",
+};
+
+export const tox: Skill = {
+  name: "Tox",
+  category: "Project Managers",
+  skillType: "hard",
+  slug: "tox",
+};
+
+export const nxjs: Skill = {
+  name: "Nx.js",
+  category: "Project Managers",
+  skillType: "hard",
+  slug: "nx-js",
+};
+
+const technicalHardSkillsProjectManagers: Skill[] = [
+  npm,
+  yarn,
+  pnpm,
+  poetry,
+  pybuilder,
+  maven,
+  gradle,
+  tox,
+  nxjs,
+];
+
+export default technicalHardSkillsProjectManagers;
